♻️ Refactor: Organize code into presentation layer
- 📦 : Move UI-related classes (Activities and Adapters) into a `presentation` package.
- 🏷️ : Rename `App` to `PivoSdkTestApplication`.
