+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
     <uses-feature android:name="android.hardware.bluetooth_le" android:required="false" />
 
     <application
-        android:name="app.pivo.android.basicsdkdemo.App"
+        android:name="app.pivo.android.basicsdkdemo.PivoSdkTestApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -24,10 +24,10 @@
         android:theme="@style/AppTheme">
 
         <activity
-            android:name="app.pivo.android.basicsdkdemo.PivoControllerActivity"/>
+            android:name="app.pivo.android.basicsdkdemo.presentation.pivocontrol.PivoControllerActivity"/>
 
         <activity
-            android:name="app.pivo.android.basicsdkdemo.PivoScanningActivity"
+            android:name="app.pivo.android.basicsdkdemo.presentation.scan.PivoScanningActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
