+++ app/src/main/java/app/pivo/android/basicsdkdemo/presentation/scan/ScanResultAdapter.kt
@@ -1,4 +1,4 @@
-package app.pivo.android.basicsdkdemo
+package app.pivo.android.basicsdkdemo.presentation.scan
 
 import android.util.Log
 import android.view.LayoutInflater
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import app.pivo.android.basicsdkdemo.R
 import app.pivo.android.sdk.model.PivoDevice
 import java.util.*
