+++ app/src/main/res/layout/activity_pivo_scanning.xml
@@ -3,8 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="app.pivo.android.basicsdkdemo.PivoScanningActivity">
+    android:layout_height="match_parent">
 
     <TextView
         android:layout_marginTop="40dp"
